requirements.txt
aiofiles==0.6.0
baidu-aip==2.2.18.0
certifi==2020.12.5
chardet==4.0.0
h11==0.9.0
httpcore==0.11.1
httptools==0.1.1
httpx==0.15.4
idna==2.10
multidict==5.0.0
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pygame==2.0.0
requests==2.25.1
rfc3986==1.4.0
sanic==20.9.1
sniffio==1.2.0
ujson==4.0.1
urllib3==1.26.2
uvloop==0.14.0
websockets==8.1
